backend/requirements.txt
Django~=5.0.2
coverage~=6.3
django-cors-headers~=3.14.0
djangorestframework-simplejwt~=5.2.2
djangorestframework~=3.14.0
drf-yasg==1.21.7
flake8==7.0.0
psycopg2-binary~=2.9.5
pytz~=2022.7.1
pyyaml==6.0.1
uritemplate==4.1.1
django_microsoft_auth==3.0.1
